requirements.txt
ansitable==0.11.4
cfgv==3.5.0
colored==2.3.1
contourpy==1.3.2
cycler==0.12.1
distlib==0.4.0
filelock==3.20.2
fonttools==4.61.1
identify==2.6.15
kiwisolver==1.4.9
matplotlib==3.7.3
mpmath==1.3.0
nodeenv==1.10.0
numpy==1.26.4
packaging==25.0
pgraph-python==0.6.3
pillow==12.1.0
platformdirs==4.5.1
pre_commit==4.5.1
progress==1.6.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
roboticstoolbox-python==1.1.1
rtb-data==1.0.1
scipy==1.15.3
six==1.17.0
spatialgeometry==1.1.0
spatialmath-python==1.1.15
swift-sim==1.1.0
sympy==1.14.0
typing_extensions==4.15.0
virtualenv==20.35.4
websockets==15.0.1
